refactor(shared): moved extension code inside 'chrome' folder to proper locations
