+++ packages/shared/src/extension/background/README.md
