+++ packages/shared/src/extension/background/reloadExtension.ts
@@ -1,5 +1,5 @@
-import config from '../../config';
-import { bo } from '../../utils/browser.utils';
+import config from '../config';
+import { bo } from '../utils/browser.utils';
 
 export const load = (): void => {
   if (config.DEVELOPMENT) {
