+++ packages/shared/src/extension/popup/index.tsx
@@ -1,16 +1,22 @@
-import '../../../../public/font.css';
+import '../../../public/font.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createTheme, ThemeProvider } from '@material-ui/core/styles';
-import Popup, { PopupProps } from './components/popup';
+import Popup, { PopupProps } from './components/Popup';
 
+// define @mui theme for popup
 const theme = createTheme({
   typography: {
     fontFamily: 'Trex-Regular',
   },
 });
 
-export function main(props: PopupProps): void {
+/**
+ * This function is responsible to render the extension popup in the proper DOM node.
+ *
+ * Each extension has to use this by giving platform specific configuration
+ */
+export function renderPopup(props: PopupProps): void {
   ReactDOM.render(
     <ThemeProvider theme={theme}>
       <Popup {...props} />
