+++ platforms/tktrex/extension/src/background/index.ts
@@ -1,6 +1,6 @@
 // the background doesn't need to be invoked,
 // so the import is sufficient
-import { load } from '@shared/extension/chrome/background/index';
+import { load } from '@shared/extension/background/index';
 import api, { getHeadersForDataDonation } from './api';
 
 load({
