+++ platforms/yttrex/extension/__tests__/app.spec.ts
@@ -1,15 +1,17 @@
+// mocks
+import { getChromeMock } from '@shared/extension/__mocks__/chrome.mock';
 import fetchMock from 'jest-fetch-mock';
 
+// imports
 import { HandshakeActiveResponseArb } from '@shared/arbitraries/HandshakeResponse.arb';
 import { boot, BootOpts } from '@shared/extension/app';
-import { load } from '@shared/extension/chrome/background/index';
+import { load } from '@shared/extension/background/index';
 import { fc } from '@shared/test';
 import axiosMock from '@shared/test/__mocks__/axios.mock';
 import { sleep } from '@shared/utils/promise.utils';
 import { youtubeDomainRegExp } from '@yttrex/shared/parsers/index';
 import * as fs from 'fs';
 import * as path from 'path';
-import { getChromeMock } from '../../../../packages/shared/src/extension/__mocks__/chrome';
 import * as app from '../src/app/app';
 import api, { getHeadersForDataDonation } from '../src/background/api';
 import * as events from '../src/handlers/events';
